infra/cdk/requirements.txt
aws-cdk-lib==2.*
constructs>=10.0.0,<11.0.0
pytest
pytest-cov
